# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/internal/admin/ledger/rejections/{id}/resolve" : {
      "post" : {
        "tags" : [ "admin-ledger-rejection-controller" ],
        "operationId" : "resolve",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LedgerRejectionDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/ledger/transactions" : {
      "get" : {
        "tags" : [ "admin-ledger-controller" ],
        "operationId" : "getTransactions",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "transactionId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "ownerId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "ownerType",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "GATEWAY_RECEIVABLE", "CASH_RECEIVABLE", "BANK", "PLATFORM_CLEARING", "PLATFORM_REVENUE", "TAX_PAYABLE", "PAYOUT_IN_TRANSIT", "RESTAURANT_PAYABLE", "DRIVER_PAYABLE", "CUSTOMER_CREDIT", "ADVERTISER_PREPAID" ]
          }
        }, {
          "name" : "category",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "DELIVERY_FEE", "PLATFORM_FIXED_FEE", "PLATFORM_BONUS", "FOOD_COST", "SGST", "CGST", "REFUND", "ORDER_TOTAL", "AD_IMPRESSION", "AD_CLICK", "AD_CONVERSION", "AD_WALLET_TOPUP", "CLAWBACK", "PAYOUT_TRANSFER", "CASH_COLLECTED", "CASH_REMITTED", "STORE_CREDIT" ]
          }
        }, {
          "name" : "direction",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "CREDIT", "DEBIT" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseDtoLedgerTransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/ledger/rejections" : {
      "get" : {
        "tags" : [ "admin-ledger-rejection-controller" ],
        "operationId" : "list",
        "parameters" : [ {
          "name" : "resolved",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        }, {
          "name" : "producer",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoLedgerRejectionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/ledger/rejections/{id}" : {
      "get" : {
        "tags" : [ "admin-ledger-rejection-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LedgerRejectionDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/ledger/entries" : {
      "get" : {
        "tags" : [ "admin-ledger-controller" ],
        "operationId" : "getEntries",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "transactionId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "ownerId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "ownerType",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "GATEWAY_RECEIVABLE", "CASH_RECEIVABLE", "BANK", "PLATFORM_CLEARING", "PLATFORM_REVENUE", "TAX_PAYABLE", "PAYOUT_IN_TRANSIT", "RESTAURANT_PAYABLE", "DRIVER_PAYABLE", "CUSTOMER_CREDIT", "ADVERTISER_PREPAID" ]
          }
        }, {
          "name" : "category",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "DELIVERY_FEE", "PLATFORM_FIXED_FEE", "PLATFORM_BONUS", "FOOD_COST", "SGST", "CGST", "REFUND", "ORDER_TOTAL", "AD_IMPRESSION", "AD_CLICK", "AD_CONVERSION", "AD_WALLET_TOPUP", "CLAWBACK", "PAYOUT_TRANSFER", "CASH_COLLECTED", "CASH_REMITTED", "STORE_CREDIT" ]
          }
        }, {
          "name" : "direction",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "CREDIT", "DEBIT" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseDtoLedgerEntry"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ResolveRequest" : {
        "type" : "object",
        "properties" : {
          "note" : {
            "type" : "string"
          }
        }
      },
      "LedgerRejectionDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "eventId" : {
            "type" : "string"
          },
          "producer" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "payload" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "resolvedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "resolvedBy" : {
            "type" : "string"
          },
          "resolutionNote" : {
            "type" : "string"
          },
          "ageMinutes" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "ApiResponsePageResponseDtoLedgerTransactionDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResponseDtoLedgerTransactionDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "LedgerTransactionDto" : {
        "required" : [ "amount", "category", "date", "fromAccountId", "toAccountId", "transactionId" ],
        "type" : "object",
        "properties" : {
          "transactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "category" : {
            "type" : "string",
            "enum" : [ "DELIVERY_FEE", "PLATFORM_FIXED_FEE", "PLATFORM_BONUS", "FOOD_COST", "SGST", "CGST", "REFUND", "ORDER_TOTAL", "AD_IMPRESSION", "AD_CLICK", "AD_CONVERSION", "AD_WALLET_TOPUP", "CLAWBACK", "PAYOUT_TRANSFER", "CASH_COLLECTED", "CASH_REMITTED", "STORE_CREDIT" ]
          },
          "fromAccountId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "toAccountId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "type" : "number"
          },
          "date" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageResponseDtoLedgerTransactionDto" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LedgerTransactionDto"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "PageResponseDtoLedgerRejectionDto" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LedgerRejectionDto"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponsePageResponseDtoLedgerEntry" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResponseDtoLedgerEntry"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "LedgerEntry" : {
        "required" : [ "accountId", "amount", "category", "createdAt", "direction", "id", "transactionId" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "transactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "referenceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "accountId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "direction" : {
            "type" : "string",
            "enum" : [ "CREDIT", "DEBIT" ]
          },
          "category" : {
            "type" : "string",
            "enum" : [ "DELIVERY_FEE", "PLATFORM_FIXED_FEE", "PLATFORM_BONUS", "FOOD_COST", "SGST", "CGST", "REFUND", "ORDER_TOTAL", "AD_IMPRESSION", "AD_CLICK", "AD_CONVERSION", "AD_WALLET_TOPUP", "CLAWBACK", "PAYOUT_TRANSFER", "CASH_COLLECTED", "CASH_REMITTED", "STORE_CREDIT" ]
          },
          "amount" : {
            "type" : "number"
          },
          "producer" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "authorizedBy" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageResponseDtoLedgerEntry" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LedgerEntry"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}